refactor(schematics): merge ast-utils and route-utils into ast-utils
There was cyclic dependency between those files. They both do AST manipulation, so I merged
both into ast-utils.
